backend/requirements.txt
fastapi[standard]>=0.115.0
sqlmodel>=0.0.22
passlib[bcrypt]>=1.7.4
bcrypt<4
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
